Refactor code to move generation logic to generation_utils.
